requirements.txt
Django==5.1
pytest==8.3.2
djangorestframework==3.15.2
pytest-django==4.8.0
django-extensions==3.2.3
ipython==8.26.0
python-dotenv==1.0.1
boto3==1.35.11